qca: ATH 12.5.CS: Add qca feeds related migration changes

+++ qca/feeds/mac80211-qca/files-qca/files/lib/apsta_mode.sh
@@ -22,7 +22,6 @@ dfs_log=1
 phy="$4"
 ml_link=""
 ap_ht_capab=$(cat $hostapd_conf 2> /dev/null | grep ht_capab | grep -v vht | cut -d'=' -f 2)
-ap_link_file="/tmp/ap_interface_link.txt"
 
 get_sta_freq_list() {
 	phy=$1
@@ -222,22 +221,157 @@ hostapd_ht20_mode() {
         hostapd_cli -i $ap_intf $ml_link set ht_capab $ht_capab_20 2> /dev/null
 }
 
+get_ap_ht_capab() {
+        sta_freq=$1
+        sta_radio_band_idx=$(ls /var/run/wpa_supplicant-*-updated-cfg | awk '{ print $1 }' | cut -f2 -d"-" | awk '{print substr($0,length,1)}')
+        for i in $sta_radio_band_idx
+        do
+                freq_list=$(cat /var/run/wpa_supplicant-radio${phy: -1}_band${i}-updated-cfg | grep freq_list | cut -d'=' -f 2)
+                highest_freq=$(echo "$freq_list" | awk '{print $NF}')
+                least_freq=$(echo "$freq_list" | cut -d ' ' -f1)
+                if [ "$sta_freq" -ge "$least_freq" ] && [ "$sta_freq" -le "$highest_freq" ]; then
+			hostapd_conf="/var/run/hostapd-${phy}_band$i.conf"
+			ap_ht_capab=$(cat $hostapd_conf 2> /dev/null | grep ht_capab | grep -v vht | cut -d'=' -f 2)
+                fi
+        done
+}
+
+hostapd_set_op_class() {
+    local freq=$1
+    local channel=$2
+    local width=$3
+    local intf=$4
+    local sec_channel=0
+    local vht_he_opclass=0
+
+    centre_freq=$(iw dev $intf info | grep $freq | awk '{print $9}')
+
+    # Determine sec_channel based on ap_ht_mode if width is 40
+    if [ "$width" -eq 40 ]; then
+	    if [ "$centre_freq" -gt "$freq" ]; then
+		    sec_channel=1
+	    elif [ "$centre_freq" -lt "$freq" ]; then
+		    sec_channel=-1
+	    else
+		    sec_channel=0
+	    fi
+    fi
+
+    # 2.4 GHz frequency range
+    if [ "$freq" -ge 2412 ] && [ "$freq" -le 2484 ]; then
+        if [ "$width" -eq 40 ]; then
+            case "$sec_channel" in
+                1) op_class=83 ;;
+                -1) op_class=84 ;;
+                *) op_class=81 ;;
+            esac
+        elif [ "$width" -eq 20 ]; then
+            if [ "$freq" -eq 2484 ]; then
+                op_class=82
+            else
+                op_class=81
+            fi
+        fi
+    fi
+
+    # 5 GHz frequency range
+    if [ "$freq" -ge 5180 ] && [ "$freq" -le 5885 ]; then
+        case "$width" in
+            80) vht_he_opclass=128 ;;  # 80 MHz
+            160) vht_he_opclass=129 ;;  # 160 MHz
+            "80+80") vht_he_opclass=130 ;;  # 80+80 MHz
+            320) vht_he_opclass=129 ;;  # 320 MHz
+            *)
+        esac
+
+        if [ "$freq" -ge 5180 ] && [ "$freq" -le 5240 ]; then
+            if [ "$vht_he_opclass" -ne 0 ]; then
+                op_class=$vht_he_opclass
+            else
+                case "$sec_channel" in
+                    1) op_class=116 ;;
+                    -1) op_class=117 ;;
+                    *) op_class=115 ;;
+                esac
+            fi
+        elif [ "$freq" -ge 5260 ] && [ "$freq" -le 5320 ]; then
+            if [ "$vht_he_opclass" -ne 0 ]; then
+                op_class=$vht_he_opclass
+            else
+                case "$sec_channel" in
+                    1) op_class=119 ;;
+                    -1) op_class=120 ;;
+                    *) op_class=118 ;;
+                esac
+            fi
+        elif [ "$freq" -ge 5745 ] && [ "$freq" -le 5885 ]; then
+            if [ "$vht_he_opclass" -ne 0 ]; then
+                op_class=$vht_he_opclass
+            else
+                case "$sec_channel" in
+                    1) op_class=126 ;;
+                    -1) op_class=127 ;;
+                    *) op_class=125 ;;
+                esac
+            fi
+        elif [ "$freq" -ge 5500 ] && [ "$freq" -le 5720 ]; then
+            if [ "$vht_he_opclass" -ne 0 ] || [ "$width" -eq 320 ]; then
+                op_class=$vht_he_opclass
+            else
+                case "$sec_channel" in
+                    1) op_class=122 ;;
+                    -1) op_class=123 ;;
+                    *) op_class=121 ;;
+                esac
+            fi
+        fi
+
+    fi
+
+    # 6 GHz frequency range
+    if [ "$freq" -ge 5950 ] && [ "$freq" -le 7115 ]; then
+        case "$width" in
+            80) op_class=133 ;;  # 80 MHz
+            160) op_class=134 ;;  # 160 MHz
+            "80+80") op_class=135 ;;  # 80+80 MHz
+            320) op_class=137 ;;  # 320 MHz
+            *)
+		    if [ "$sec_channel" -ne 0 ]; then
+			    op_class=132
+		    else
+			    op_class=131
+		    fi
+		    ;;
+        esac
+    fi
+
+    if [ "$freq" -eq 5950 ];then
+	    op_class=136
+    fi
+    if [ "$freq" -ge 5500 ] && [ "$freq" -le 5720 ]; then
+	    if [ "$width" -eq 320 ]; then
+		return
+	    fi
+    fi
+
+	hostapd_cli -i $ap_intf $ml_link set op_class $op_class> /dev/null
+}
+
 # STA association is completed, hence adjusting hostapd running config
 hostapd_adjust_config() {
-	sta_freq=$1
-	sta_channel=$(iw $phy channels | grep $sta_freq |  awk '{print $4}' |  sed -e "s/\[//g" | sed -e "s/\]//g")
-	sta_width=$(iw dev $sta_intf info | grep $sta_freq | awk '{print $6}')
-	wifi_gen=$(wpa_cli -i $sta_intf status 2> /dev/null | grep wifi_generation | cut -d'=' -f 2)
-	ieee80211ac=$(wpa_cli -i $sta_intf status 2> /dev/null | grep ieee80211ac | cut -d'=' -f 2)
-	ap_intf=$2
-	wifi_6gband=$(hostapd_is_6ghz_band $sta_freq)
-	wifi_5gband=$(hostapd_is_5ghz_band $sta_freq)
+        sta_freq=$1
+        sta_channel=$(iw $phy channels | grep $sta_freq |  awk '{print $4}' |  sed -e "s/\[//g" | sed -e "s/\]//g")
+        sta_width=$(iw dev $sta_intf info | grep $sta_freq | awk '{print $6}')
+        wifi_gen=$(wpa_cli -i $sta_intf status 2> /dev/null | grep wifi_generation | cut -d'=' -f 2)
+        ieee80211ac=$(wpa_cli -i $sta_intf status 2> /dev/null | grep ieee80211ac | cut -d'=' -f 2)
+        ap_intf=$2
+        wifi_6gband=$(hostapd_is_6ghz_band $sta_freq)
+        wifi_5gband=$(hostapd_is_5ghz_band $sta_freq)
 
 	if [ -z $ieee80211ac ]; then
 		ieee80211ac=0
 	fi
 
-	echo "STA associated in Channel $sta_channel, Width $sta_width MHz, Wifi Gen $wifi_gen, AP $ap_intf link $ml_link" > /dev/ttyMSM0
 
 	hostapd_cli -i $ap_intf $ml_link set channel $sta_channel 2> /dev/null
 	if [ "$wifi_5gband" == "true" ] || [ "$wifi_6gband" == "true" ]; then
@@ -245,6 +379,8 @@ hostapd_adjust_config() {
 	else
 		hostapd_cli -i $ap_intf $ml_link set hw_mode g 2> /dev/null
 	fi
+
+	get_ap_ht_capab $sta_freq
 
 	ap_ht_mode=$(echo $ap_ht_capab | sed -n 's/.*\(\[HT40*+*-*]\).*/\1/p')
 	#echo "Current AP HT capab $ap_ht_capab" > /dev/ttyMSM0
@@ -345,29 +481,22 @@ hostapd_adjust_config() {
                         hostapd_ht40_mode "$sta_channel"
                 fi
 	fi
+	hostapd_set_op_class $sta_freq $sta_channel $sta_width $sta_intf
+	echo "STA associated in Channel $sta_channel, Width $sta_width MHz, Wifi Gen $wifi_gen, AP $ap_intf link $ml_link op_class $op_class" > /dev/ttyMSM0
 }
 
 get_link_ids() {
 	ifname=$1
-	link=$(iw dev $ifname info | grep link | cut -d':' -f 1 2> /dev/null  | cut -d ' ' -f 2)
 
-	if [ -n "$link" ]; then
-		echo "$link"
-		return
-	else
-		#when interface is in disable state, we might not find link from iw, hence, use control
-		#interface file names to find the link id
-
-		ctrl_iface=$(ls /var/run/hostapd/${ifname}*)
-		if [ -n "$ctrl_iface" ]; then
-			def_ctrl_iface_path=$(ls /var/run/hostapd/$ifname* | head -n 1)
-			#Try to return links only if has link control interface
-			if [[ "$def_ctrl_iface_path" == *"link"* ]]; then
-				links=$(ls /var/run/hostapd/$ifname* | awk '{print substr($0,length,1)}')
-				if [ -n "$links" ]; then
-					echo "$links"
-					return
-				fi
+	ctrl_iface=$(ls /var/run/hostapd/${ifname}_*)
+	if [ -n "$ctrl_iface" ]; then
+		def_ctrl_iface_path=$(ls /var/run/hostapd/${ifname}_* | head -n 1)
+		#Try to return links only if has link control interface
+		if [[ "$def_ctrl_iface_path" == *"link"* ]]; then
+			links=$(ls /var/run/hostapd/${ifname}_* | awk '{print substr($0,length,1)}')
+			if [ -n "$links" ]; then
+				echo "$links"
+				return
 			fi
 		fi
 	fi
@@ -480,10 +609,7 @@ EOF
 chmod 777 /lib/radar_detect.sh
 wpa_cli -i $sta_intf -a /lib/radar_detect.sh &
 
-if [ -f $ap_link_file]; then
-	rm $ap_link_file
-fi
-#echo "Checking wpa_state $(wpa_cli -i $sta_intf status 2> /dev/null | grep wpa_state | cut -d'=' -f 2)" > /dev/ttyMSM0
+echo "Checking wpa_state $(wpa_cli -i $sta_intf status 2> /dev/null | grep wpa_state | cut -d'=' -f 2)" > /dev/ttyMSM0
 
 while true;
 do
@@ -525,8 +651,6 @@ do
 				else
 					hostapd_cli -i $ap_intf disable
 				fi
-				#echo "Last link disabled: $i" > /dev/ttyMSM0
-				echo "$ap_intf=$i" >> $ap_link_file
 			fi
 		done
 	fi
@@ -548,7 +672,7 @@ do
 				for ap_intf in $ap_intfs
 				do
 					ml_link=$(get_link_info $ap_intf $freq)
-					#echo link config command is $ml_link $freq $ap_intf  > /dev/console
+					echo link config command is $ml_link $freq $ap_intf  > /dev/console
 					if [ -n "$ml_link" ]; then
 						hostapd_adjust_config $freq $ap_intf
 					fi
@@ -561,7 +685,7 @@ do
 			# workaround for upstream station mld failed to get sta freq list
 			if [ $sta_freq -eq 0 ] && [ $wifi_gen -eq 6 ]; then
 				sta_freq=$(wpa_cli -i $sta_intf mlo_status 2> /dev/null | grep freq | cut -d'=' -f 2)
-			#	echo "sta freq $sta_freq" >> /tmp/apsta_debug.log
+				echo "sta freq $sta_freq" >> /tmp/apsta_debug.log
 			fi
 
 			wifi_6gband=$(hostapd_is_6ghz_band $sta_freq)
@@ -579,30 +703,14 @@ do
 
 		ap_status=$(hostapd_get_ap_status $ap_intf)
 
-		if [ "$ap_status" = "DISABLED" ]; then
+		if [ $(wpa_cli -i $sta_intf status 2> /dev/null | grep wpa_state | cut -d'=' -f 2) = "COMPLETED" ] &&
+		   [ "$ap_status" = "DISABLED" ]; then
 			for ap_intf in $ap_intfs
 			do
 				ap_links=$(get_link_ids $ap_intf)
 				if [ -n "$ap_links" ]; then
-
-					if [ -f $ap_link_file ]; then
-						#This is needed temporarily as hostapd needs to disable order
-						#to be restored atleast for the last disabled link
-						while read link_file; do
-							last_ap_intf=`echo $link_file | cut -d'=' -f1`
-							if [ $last_ap_intf = $ap_intf ];then
-								mld_disable_last_link=`echo $link_file | cut -d'=' -f2`
-								break;
-							fi
-						done < $ap_link_file
-					fi
-					#echo "Last link going to enable: $mld_disable_last_link" > /dev/ttyMSM0
-					hostapd_cli -i $ap_intf -l $mld_disable_last_link enable
 					for i in $ap_links
 					do
-						if [ $i = $mld_disable_last_link ];then
-							continue;
-						fi
 						hostapd_cli -i $ap_intf -l $i enable
 					done
 				else
@@ -611,27 +719,13 @@ do
 
 				ap_status=$(hostapd_get_ap_status $ap_intf)
 				# workaround for single instance hostapd not doing "enable" without "disable" call to deinit hapd driver
-				if [ $ap_status = "DISABLED" ]; then
-
+				if [ $(wpa_cli -i $sta_intf status 2> /dev/null | grep wpa_state | cut -d'=' -f 2) = "COMPLETED" ] &&
+				   [ "$ap_status" = "DISABLED" ]; then
 					if [ -n "$ap_links" ]; then
 						for i in $ap_links
 						do
 							hostapd_cli -i $ap_intf -l $i disable
-							sleep 5
-						done
-
-						#This is needed temporarily as hostapd needs to disable order
-						#to be restored atleast for the last disabled link
-
-						mld_disable_last_link=$i
-						#echo "Last link going to enable: $mld_disable_last_link" > /dev/ttyMSM0
-						hostapd_cli -i $ap_intf -l $mld_disable_last_link enable
-
-						for i in $ap_links
-						do
-							if [ $i = $mld_disable_last_link ];then
-								continue;
-							fi
+							sleep 1
 							hostapd_cli -i $ap_intf -l $i enable
 							sleep 4
 						done
@@ -643,17 +737,16 @@ do
 					fi
 				fi
 
-				if [ -f $ap_link_file ]; then
-					rm $ap_link_file
-				fi
-
 				ap_status=$(hostapd_get_ap_status $ap_intf)
-				if [ "$ap_status" = "DISABLED" -o "$ap_status" = "FAIL" ]; then
-					echo "REPEATER AP $ap_intf failed bring-up, exiting" > /dev/ttyMSM0
+				if [ "$ap_status" = "DISABLED" -o "$ap_status" = "FAIL" ] &&
+				   [ $(wpa_cli -i $sta_intf status 2> /dev/null | grep wpa_state | cut -d'=' -f 2) = "COMPLETED" ]; then
+					echo "REPEATER AP $ap_intf failed bring-up, status $ap_status exiting" > /dev/ttyMSM0
 					logread > /tmp/logread_AP_failure.log
 					echo "Collect if any core present in /tmp/ and output of /tmp/logread_AP_failure.log" > /dev/console
 					echo "Hostapd enable failed, exiting" >> /tmp/apsta_debug.log
 					date >> /tmp/apsta_debug.log
+					iw dev >> /tmp/apsta_debug.log
+					iw dev $ap_intf info >> /tmp/apsta_debug.log
 					ap_link=$(get_link_ids $ap_intf)
 					if [ -n "$ap_link" ]; then
 						for i in $ap_link
